requirements.txt
asgiref==3.2.7
astroid==2.4.0
colorama==0.4.3
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
PyJWT==1.7.1
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
wrapt==1.12.1
